% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoDiatomRiverData.R
\docType{data}
\name{demoDiatomRiverData}
\alias{demoDiatomRiverData}
\title{This is demo data to be included in my package}
\description{
This is demo data to be included in my package
}
\references{
\url{www.sepa.org.uk}
}
\keyword{data}
