% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquaMetrics.R
\docType{data}
\name{LookUpDARLEQ2}
\alias{LookUpDARLEQ2}
\title{DARLEQ. TDIs associated with the presence of diatoms.}
\format{A data frame with 1275 rows and 24 variables}
\description{
TDI (Trophic Diatom Index) associated with diatoms in rivers and lakes,
providing a information on the nutrient status of rivers, as appear in DARLEQ2
}
\details{
The columns are as follows:

\itemize{
 \item TaxonId.
 \item NBSCode.
 \item AccCode.
 \item Class.
 \item Family.
 \item Authority.
 \item Genus.
 \item TaxonName.
 \item DARESTaxonName.
 \item DARESAuthority.
 \item Aggregate.
 \item Status.
 \item TDIndices (TDIo, TDI3, TDI4, LTDI1, LTDI2, LTDI3)
 \item DAM.
 \item Motile.
 \item OrganicTolerant.
 \item Planktic.
 \item Saline.
 \item Notes.
}
}
\keyword{datasets}
